DFDF-9454 review fixes
Former-commit-id: 55d2df3e84270016209b6255af9659a79fbd9dac

+++ Sources/Internal/Math/MathHelpers.h
@@ -189,15 +189,15 @@ inline bool GetIntersectionVectorWithPlane(const Vector3& start, const Vector3& 
 	*/
 inline float32 SquareRootFloat(float32 number)
 {
-    long i;
+    int32 i;
     float32 x, y;
     const float32 f = 1.5f;
 
     x = number * 0.5f;
     y = number;
-    i = *(reinterpret_cast<long*>(&y));
+    i = *(reinterpret_cast<int32*>(&y));
     i = 0x5f3759df - (i >> 1);
-    y = *(reinterpret_cast<float*>(&i));
+    y = *(reinterpret_cast<float32*>(&i));
     y = y * (f - (x * y * y));
     y = y * (f - (x * y * y));
     return number * y;
